+++ utilities/findcp/findcp.f90
@@ -120,7 +120,8 @@ subroutine calcHess(natoms,cgeom,nstate,istate,stepsize,hessian,skip)
   double precision,intent(out) :: hessian(3*natoms,3*natoms)
   double precision   ::  mdif
   
-  integer   ::   i,  j, skipdisp(natoms*3)
+  integer   ::   i,  j
+  logical   ::   skipdisp(natoms*3)
   double precision  ::  dispgeom(3*natoms), dgrd(3*natoms)
   real*8    ::  h(nstate,nstate),cg(3*natoms,nstate,nstate),dcg(3*natoms,nstate,nstate),e(nstate)
   skipdisp=.false.
@@ -168,7 +169,7 @@ subroutine findmin(natoms,nstate,cgeom,isurf,maxiter,shift,Etol,Stol)
   double precision            :: nrmG, nrmD, tmp(1)
   double precision, external  :: dnrm2
   double precision,  parameter  :: amu2au=1.822888484514D3,au2cm1=219474.6305d0
-  double precision, parameter   :: MAXD = 1D-2
+  double precision, parameter   :: MAXD = 1D-1
 
   ! initialize work spaces
   call DSYEVD('V','U',3*natoms,hess,3*natoms,w,tmp,-1,itmp,-1,INFO)
@@ -179,6 +180,8 @@ subroutine findmin(natoms,nstate,cgeom,isurf,maxiter,shift,Etol,Stol)
   allocate(IWORK(LIWORK))
 
   print "(A,I4,A,I4,A)","Searching for minimum on surface ",isurf," in ",maxiter," iterations."
+  print "(A)","  Convergence Tolerances"
+  print "(A,E10.2,A,E10.2)","  Energy Gradient: ",Etol,"   Displacement:",Stol
   do iter=1,maxiter
      call EvaluateSurfgen(cgeom,e,cg,h,dcg)
      grad = cg(:,isurf,isurf)
@@ -227,140 +230,80 @@ end module opttools
 program findcp
 
  use opttools 
-implicit none
+ implicit none
 
-  integer,parameter :: natm = 13, nst=4
-! gmin is the global min
-  real*8  :: gmin(39)=(/	-1.71199414,	0.00000000,	-0.00354718,	&
-				-0.38703243,	0.00000000,	-2.28591809,	&
-				2.25530537,	0.00000000,	-2.27653197,	&
-				3.58567656,	0.00000000,	-0.00071621,	&
-				2.24421854,	0.00000000,	2.27787307,	&
-				-0.39002962,	0.00000000,	2.28562351 ,	&
-				-4.25258741,	0.00000000,	0.10849745,	&
-				-1.40667035,	0.00000000,	-4.06095466,	&
-				3.26109586,	0.00000000,	-4.05562411,	&
-				5.62831685,	0.00000000,	0.00217672,	&
-				3.24992294,	0.00000000,	4.05714529,	&
-				-1.44534888,	0.00000000,	4.03356585,	&
-				-4.94259551,	0.00000000,	-1.58803693 	/)
+  integer :: natm , nst, ngeoms
+  integer :: i,j, isurf, ios
+  double precision,allocatable  :: anum(:),masses(:),  hess(:,:), w(:)
+  character*3,allocatable       :: aname(:)
+  double precision,allocatable  :: cgeom(:),hvec(:)
+  logical,allocatable           :: skip(:)
+  double precision,allocatable  :: h(:,:),cg(:,:,:),dcg(:,:,:),e(:),evec(:,:)
+  double precision,external     :: dnrm2
+  character*300                 ::  geomfile,str
 
-  real*8  :: dis(39)=(/	-2.18996537,	0.00000000,	0.00000000,	&
-			-0.74939941,	0.00000000,	2.34677175,	&
-			1.85640018,	0.00000000,	2.31756577,	&
-			3.18392345,	0.00000000,	0.00000000,	&
-			1.85640018,	0.00000000,	-2.31756577,	&
-			-0.74939941,	0.00000000,	-2.34677175,	&
-			-4.51954534,	0.00000000,	0.00000000,	&
-			-1.80877667,	0.00000000,	4.09382748,	&
-			2.90247794,	0.00000000,	4.07428219,	&
-			5.22705676,	0.00000000,	0.00000000,	&
-			2.90247794,	0.00000000,	-4.07428219,	&
-			-1.80877667,	0.00000000,	-4.09382748,	&
-			30.00000000,	0.00000000,	0.00000000	/) 
-! the minimum energy intersection
-  real*8  :: mex01(39) =(/	-1.74652425,	0.00000000,	0.07042240,	&
-				-0.39925857,	0.00000000,	-2.23489349,	&
-				2.22489298,	0.00000000,	-2.26280385,	&
-				3.59825936,	0.00000000,	-0.00272005,	&
-				2.29561690,	0.00000000,	2.30445676,	&
-				-0.33931315,	0.00000000,	2.37489408,	&
-				-4.21991527,	0.00000000,	0.09775378,	&
-				-1.47344284,	0.00000000,	-3.96658552,	&
-				3.19890493,	0.00000000,	-4.05833294,	&
-				5.64292066,	0.00000000,	-0.04200458,	&
-				3.36128528,	0.00000000,	4.05356284,	&
-				-1.30715914,	0.00000000,	4.17552016,	&
-				-5.73132025,	0.00000000,	-3.00077829	/)
-  real*8  :: mex12(39) =(/	-1.71136176,	0.00000000,	-0.00965084,	&
-				-0.36782782,	0.00000000,	-2.36235648,	&
-				2.22626539,	0.00000000,	-2.32736351,	&
-				3.55065740,	0.00000000,	0.00664434,	&
-				2.22568282,	0.00000000,	2.33210381,	&
-				-0.37264378,	0.00000000,	2.35055752,	&
-				-4.11963171,	0.00000000,	0.11212956,	&
-				-1.42605725,	0.00000000,	-4.10703334,	&
-				3.26775400,	0.00000000,	-4.08237091,	&
-				5.59261842,	0.00000000,	0.00434343,	&
-				3.26037605,	0.00000000,	4.08939387,	&
-				-1.46705199,	0.00000000,	4.07053107,	&
-				-4.95464207,	0.00000000,	-1.57906004	/)
 
-  double precision,parameter :: au2cm1=219474.6305d0
-  double precision, parameter  ::  bohr2ang=0.529177249d0
-
-  double precision  ::  masses(natm),  hess(natm*3,natm*3), w(3*natm)
-  double precision  ::  cgeom(3*natm),hvec(3*natm)
-  integer  :: i,j, isurf
-  logical :: skip(natm)= .false.
-  double precision  ::  di, dj, x,y, r, r0, dr,dh
-  real*8    :: h(nst,nst),cg(3*natm,nst,nst),dcg(3*natm,nst,nst),e(nst),evec(nst,nst)
-  double precision,dimension(3)::d1,d2,d3,cpd
-  double precision,external :: dnrm2
-  character*9,dimension(4)  ::  fnames
-
+  print *," ***************************************** "
+  print *," *    findcp.x                           * "
+  print *," ***************************************** "
+  print *," Critical point search on Surfgen surface" 
   call initPotential()
+  call getInfo(natm,nst)
 
-! Plot branching space
-  fnames(1)  = "surf1.csv"
-  fnames(2)  = "surf2.csv"
-  fnames(3)  = "surf3.csv"
-  fnames(4)  = "surf4.csv"
-  d1 = mex12(37:39)-mex12(19:21)
-  r0 = dnrm2(3,d1,1)
-  d1 = d1/r0
-  dr = 6d-2
-  dh = 1d-2
-  print *,"r0=",r0*bohr2ang
-  do isurf=1,4
-   open(unit=123+isurf,file=fnames(isurf),access='sequential',action='write')
-   do i=-5 ,40 
-      cgeom = mex12
-      r = r0+i*dr
-      cgeom(37:39) = cgeom(19:21)+r*d1
-      call EvaluateSurfgen(cgeom,e,cg,h,dcg)
-      call getEvec(evec)
-      write(unit=123+isurf,fmt=1000),r*bohr2ang,0d0,e(isurf)*au2cm1,evec(:,isurf)
-      hvec = dcg(:,2,3)
-      hvec = hvec / dnrm2(3*natm,hvec,1)
-      do j=1,35
-        cgeom = cgeom+hvec*dh
-        call EvaluateSurfgen(cgeom,e,cg,h,dcg)
-        call getEvec(evec)
-        write(unit=123+isurf,fmt=1000),r*bohr2ang,dh*j,e(isurf)*au2cm1,evec(:,isurf)
-      end do
-   end do!i
-   close(123+isurf)
-  end do
+! allocate arrays
 
-stop
+  print "(A,I6)","Number of Atoms:  ",natm
+  print "(A,I6)","Number of States: ",nst 
 
-  masses = (/	12.,12.,12.,12.,12.,12.,15.99491464,	&
-		1.00782504,1.00782504,1.00782504,1.00782504,1.00782504,1.00782504 /) 
+  allocate(masses(natm))
+  allocate(anum(natm))
+  allocate(aname(natm))
+  allocate(hess(natm*3,natm*3))
+  allocate(w(3*natm))
+  allocate(cgeom(3*natm))
+  allocate(hvec(3*natm))
+  allocate(skip(natm))
+  skip = .false.
+  allocate(h(nst,nst))
+  allocate(cg(3*natm,nst,nst))
+  allocate(dcg(3*natm,nst,nst))
+  allocate(e(nst))
+  allocate(evec(nst,nst)) 
 
-1000 format(2(F12.7,', '),F20.5,4(',',F12.8))
-  cgeom = gmin
-  isurf = 1
-!  do i=1,300
-!    call EvaluateSurfgen(cgeom,e,cg,h,dcg)
-!         d1 = cgeom(4:6)-cgeom(1:3)
-!         d2 = cgeom(7:9)-cgeom(1:3)
-!         d3 = cgeom(10:12)-cgeom(1:3)
-!         d3 = d3/dnrm2(3,d3,1)
-!         cpd(1) = d1(3)*(d2(2)-d3(2))+d2(3)*d3(2)-d2(2)*d3(3)+d1(2)*(d3(3)-d2(3))
-!         cpd(2) =-d2(3)*d3(1)+d1(3)*(d3(1)-d2(1))+d1(1)*(d2(3)-d3(3)) +d2(1)*d3(3)
-!         cpd(3) = d1(2)*(d2(1)-d3(1))+d2(2)*d3(1)-d2(1)*d3(2)+d1(1)*(d3(2)-d2(2))
-!    print "(I4,4F14.6)",i,e*au2cm1,(e(2)-e(1))*au2cm1, 90/Acos(0d0)* &
-!           asin((-d1(3)*d2(2)*d3(1)+d1(2)*d2(3)*d3(1)+d1(3)*d2(1)*d3(2)       &
-!                -d1(1)*d2(3)*d3(2)-d1(2)*d2(1)*d3(3)+d1(1)*d2(2)*d3(3))/      &
-!               dnrm2(3,cpd,1))   
-!    cgeom = cgeom-cg(:,2,2)*1D-3
-!    if((e(2)-e(1))*au2cm1<1)exit
-!  end do
+! process arguments
+! synposis:    findcp.x geom type states
+! Default values:
+! geom        geom
+! type        min
+! states      1
+  call get_command_argument(number=1,value=geomfile,status=ios)
+  if(ios<-1)  &     ! input argument larger than container
+      stop "Filename truncated.  Please use less than 300 characters"
+  if(ios>0) then
+    print *,"No filename suppied.  Using default."
+    write(geomfile,"(A)"), "geom"
+    isurf = 1
+  else
+    call get_command_argument(number=2,value=str,status=ios)
+    if(ios/=0)then
+      print *,"Cannot get surface number from command line options.  Using default."
+      isurf = 1
+    else
+      read(str,*,iostat=ios)isurf
+      if(ios/=0)then
+        print *,"Cannot get surface number from command line options.  Using default."
+        isurf = 1
+      end if
+    end if
+  end if
+
+  print *,"Reading input from input file "//trim(geomfile)
+  call readColGeom(geomfile,1,natm,aname,anum,cgeom,masses)
+
   print "(/,A)","-------------- Initial Geometry ------------"
   call analysegeom(natm,cgeom)
   print "(/,A)","----------- Geometry Optimizations ---------"
-  call findmin(natm,nst,cgeom,isurf,100,1d-2,1d-10,1d-5)
+  call findmin(natm,nst,cgeom,isurf,100,1d-3,1d-7 ,1d-5)
   print "(/,A)","--------------- Final Geometry -------------"
   call analysegeom(natm,cgeom)
   print "(/,A)","------------ Harmonic Frequencies ----------"
@@ -369,4 +312,20 @@ stop
   do i=1,3*natm
     print "(I5,F12.2)",i,w(i)
   end do
+
+! deallocate arrays
+  deallocate(masses)
+  deallocate(anum)
+  deallocate(aname)
+  deallocate(hess)
+  deallocate(w)
+  deallocate(cgeom)
+  deallocate(hvec)
+  deallocate(skip)
+  deallocate(h)
+  deallocate(cg)
+  deallocate(dcg)
+  deallocate(e)
+  deallocate(evec) 
+
 end