+++ utilities/findcp/install.sh
@@ -8,4 +8,6 @@ echo $SGENFC findcp.f90 -o findcp.x $SGENFLAG $SGENLIB
 $SGENFC findcp.f90 -o findcp.x $SGENFLAG $SGENLIB
 echo Cleaning up
 rm opttools.mod
+echo Copying executable to surfgen directory
+cp findcp.x $SGENDIR 
 echo Done