+++ utilities/ghplot/install.sh
@@ -6,4 +6,6 @@ source ../../bin/setsgenvars.sh
 echo Compiling ghplot
 echo $SGENFC ghplot.f90 -o ghplot.x $SGENFLAG $SGENLIB
 $SGENFC ghplot.f90 -o ghplot.x $SGENFLAG $SGENLIB
+echo Copying executable to surfgen directory
+cp ghplot.x $SGENDIR
 echo Done