+++ utilities/testpoints/install.sh
@@ -6,4 +6,6 @@ source ../../bin/setsgenvars.sh
 echo Compiling testpoints
 echo $SGENFC testpoints.f90 -o testpoints.x $SGENFLAG $SGENLIB
 $SGENFC testpoints.f90 -o testpoints.x $SGENFLAG $SGENLIB
+echo Copying executable to surfgen directory
+cp testpoints.x $SGENDIR
 echo Done